ci(integration): run full tests on windows
This change is modifying our CI pipeline to actually run all the
integration tests on windows. :100:
